requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
fake-useragent==1.4.0
nltk==3.8.1
textstat==0.7.3
spacy==3.7.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
Pillow==10.1.0
python-whois==0.8.0
builtwith==1.3.4
wordcloud==1.9.2
pyquery==1.4.3
gunicorn==21.2.0
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
lxml==4.9.3
urllib3==2.1.0
chromedriver-autoinstaller==0.6.4